barebox-state: get devicetree from file
Adds a -i/--input argument to barebox-state to allow passing a
devicetree as a file instead of using it from the system.
This can be used for example on systems that do not use device trees
(such as x86) but where we want to use a dtb blob for describing the
state storage and format.

Signed-off-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>
Signed-off-by: Enrico Jorns <ejo@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>