barebox-state: add new --update option to avoid writes if possible
barebox-state --set always writes the state when successful. Users
seeking to conserve write cycles thus have to --get the variable in
question first to check whether to write it. Make life of such users
easier by having barebox-state support this out-of-the-box.

Use of the new --update option will only write a variable if its
new and old value compare unequal. This allows users to fire and
forget execute barebox-state.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>