libdt: add CONFIG_TEST_LOOPBACK
We have quite a bit of tricky udev and device tree parsing code that is
only tested manually. Best case we would have tests for this in QEMU,
but until we do, let's test it with loop devices:

The downside is that we can only test block devices and that we need
a tiny bit of code that's not used normally, but on the upside, we
have tests. :-)

Tested-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.pengutronix.de/20230531153125.1408092-5-a.fatoum@pengutronix.de
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>