state: backend: direct: open block device in read-only mode if possible
We unconditionally open the device backing a direct bucket in read-write
mode. The barebox_state utility already populates struct
state_backend_storage::readonly though, which we could consult at device
open time. Do so. This could possibly be done for MTD as well, but until
that's tested, for now, we do it only for the direct backend meant for
use with block devices.

Tested-by: Ulrich Ölmann <u.oelmann@pengutronix.de>
Link: https://lore.pengutronix.de/20230607121628.3180139-2-a.fatoum@pengutronix.de
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>